use 1 folder for converting model responses
